patch_langflow.py

# This file patches potential issues with Langflow's custom types when running in standalone mode.
import uuid
from typing import Any

# Attempt to patch UUIDstr if it exists in the potential import path
# We need to make sure that when 'langflow.services.database.models.base' is imported,
# UUIDstr is compatible with SQLModel.

# One way is to pre-load the module and modify it.
import sys
from unittest.mock import MagicMock

# If langflow is not installed or available, this does nothing.
# If it IS available (via langflow-base), we might hit the bug.

try:
    # We will define a dummy UUIDstr that is just UUID
    # but we need to inject it into the module before it's used by others?
    # No, usually the issue is the definition inside the module itself.
    # We can't easily change source code of installed lib.
    # But we can monkeypatch.
    
    import langflow.services.database.models.base
    langflow.services.database.models.base.UUIDstr = uuid.UUID
except ImportError:
    pass
except Exception:
    pass
