requirements.txt
python-dotenv
lfx @ git+https://github.com/langflow-ai/langflow.git#subdirectory=src/lfx
langflow-base
langchain-core
httpx
pydantic
loguru
rich
fastapi
uvicorn
sqlalchemy
# Add your component dependencies below (e.g. openai, langchain-openai)
openai
langchain